+++ scripts/bulk-initialize-states-at-lockup.ts
@@ -1,119 +1,20 @@
-/* eslint-disable no-useless-call */
 /* eslint-disable no-undef */
-import bent from 'bent'
-import Queue from 'p-queue'
 import Web3 from 'web3'
-import {Contract} from 'web3-eth-contract/types'
+import {
+	prepare,
+	createGraphQLFetcher,
+	createEGSFetcher,
+	createGetStorageLastCumulativeGlobalReward,
+	createGetStorageLastCumulativeLockedUpAndBlock,
+	createDifferenceCaller,
+	createGetCumulativeLockedUpCaller,
+	createInitializeStatesAtLockup,
+	createQueue,
+} from './lib/bulk-initialize-states-at-lockup'
+import {graphql, ethgas} from './lib/api'
+import {GraphQLResponse} from './lib/types'
 const {CONFIG, EGS_TOKEN} = process.env
 const {log: ____log} = console
-type GraphQLResponse = {
-	data: {
-		account_lockup: Array<{
-			property_address: string
-			account_address: string
-			lockedup: {
-				block_number: number
-			}
-		}>
-	}
-}
-type EGSResponse = {
-	fast: number
-	fastest: number
-	safeLow: number
-	average: number
-	block_time: number
-	blockNum: number
-	speed: number
-	safeLowWait: number
-	avgWait: number
-	fastWait: number
-	fastestWait: number
-}
-type GasPriceFetcher = () => Promise<string>
-type SendTx = {
-	on: (event: 'confirmation' | 'error', callback: () => void) => SendTx
-}
-
-const graphql = () =>
-	bent('https://api.devprtcl.com/v1/graphql', 'POST', 'json')
-const ethgas = (token: string) =>
-	bent(`https://ethgasstation.info/api/ethgasAPI.json?api-key=${token}`, 'json')
-const prepare = async (configAddress: string) => {
-	const [config] = await Promise.all([
-		artifacts.require('AddressConfig').at(configAddress),
-	])
-	const [lockup] = await Promise.all<any>([
-		artifacts.require('Lockup').at(await config.lockup()),
-	])
-	const contract = new (web3 as Web3).eth.Contract(lockup.abi, lockup.address)
-	return contract
-}
-
-const createGraphQLFetcher = (
-	fetcher: bent.RequestFunction<bent.ValidResponse>
-) => async (offset = 0): Promise<GraphQLResponse> =>
-	fetcher('/', {
-		query: `{
-			account_lockup(
-				offset: ${offset},
-				distinct_on: [account_address, property_address],
-				order_by: [
-					{account_address: desc_nulls_last},
-					{property_address: desc_nulls_last}
-					{lockedup: {block_number: desc_nulls_last}}
-			  	]
-			) {
-				property_address
-				account_address
-				lockedup {
-					block_number
-				}
-			}
-		}`,
-	}).then((r) => (r as unknown) as GraphQLResponse)
-const createEGSFetcher = (
-	fetcher: bent.RequestFunction<bent.ValidResponse>
-) => async (): Promise<EGSResponse> =>
-	fetcher('').then((r) => (r as unknown) as EGSResponse)
-const createGetStorageLastCumulativeGlobalReward = (lockup: Contract) => (
-	blockNumber?: number
-) => async (property: string, user: string): Promise<string> =>
-	lockup.methods
-		.getStorageLastCumulativeGlobalReward(property, user)
-		.call(undefined, blockNumber)
-const createGetStorageLastCumulativeLockedUpAndBlock = (lockup: Contract) => (
-	blockNumber?: number
-) => async (
-	property: string,
-	user: string
-): Promise<{_cLocked: string; _block: string}> =>
-	lockup.methods
-		.getStorageLastCumulativeLockedUpAndBlock(property, user)
-		.call(undefined, blockNumber)
-const createDifferenceCaller = (lockup: Contract) => (
-	blockNumber?: number
-) => async (property: string): Promise<{_reward: string}> =>
-	lockup.methods.difference(property).call(undefined, blockNumber)
-const createGetCumulativeLockedUpCaller = (lockup: Contract) => (
-	blockNumber?: number
-) => async (property: string, user: string): Promise<{_value: string}> =>
-	lockup.methods
-		.getCumulativeLockedUp(property, user)
-		.call(undefined, blockNumber)
-const createInitializeStatesAtLockup = (lockup: Contract) => (from: string) => (
-	property: string,
-	user: string,
-	reward: string,
-	cLocked: string,
-	block: string,
-	gasPrice: string
-	// eslint-disable-next-line max-params
-): SendTx =>
-	lockup.methods
-		.initializeStatesAtLockup(property, user, reward, cLocked, block)
-		.send({gasPrice, from})
-const createQueue = (concurrency: number) => new Queue({concurrency})
 
 const handler = async (
 	callback: (err: Error | null) => void
@@ -154,7 +55,6 @@ const handler = async (
 		fetchGasPrice().then((res) =>
 			(web3 as Web3).utils.toWei(`${res.fastest / 10}`, 'Gwei')
 		)
-	____log(await fetchFastestGasPrice())
 
 	const lastCumulativeGlobalReward = createGetStorageLastCumulativeGlobalReward(
 		lockup
@@ -167,6 +67,8 @@ const handler = async (
 	const initializeStatesAtLockup = createInitializeStatesAtLockup(lockup)(from)
 
 	const targets = all.filter(({lockedup}) => Boolean(lockedup))
+	____log('all targets', targets.length)
+
 	const filteringTacks = targets.map(
 		({property_address, account_address, ...x}) => async () => {
 			const [cReward, {_cLocked, _block}] = await Promise.all([
@@ -174,13 +76,15 @@ const handler = async (
 				lastCumulativeLockedUpAndBlock()(property_address, account_address),
 			])
 			const skip = [cReward, _cLocked, _block].every((y) => y !== '0')
-			____log('Skip', property_address, account_address)
+			____log('Should skip item', property_address, account_address)
 			return {property_address, account_address, skip, ...x}
 		}
 	)
 	const shouldInitilizeItems = await createQueue(10)
 		.addAll(filteringTacks)
 		.then((done) => done.filter((x) => !x.skip))
+	____log('Should slip items', targets.length - shouldInitilizeItems.length)
+	____log('Should initilize items', shouldInitilizeItems.length)
 
 	const initializeTasks = shouldInitilizeItems.map(
 		({property_address, account_address, lockedup}) => async () => {
@@ -191,9 +95,16 @@ const handler = async (
 			])
 			const reward = res[0]._reward
 			const cLocked = res[1]._value
-			____log(property_address, account_address, reward, cLocked)
-
 			const gasPrice = await fetchFastestGasPrice()
+			____log(
+				'Start initilization',
+				property_address,
+				account_address,
+				reward,
+				cLocked,
+				gasPrice
+			)
+
 			await new Promise((resolve, reject) => {
 				initializeStatesAtLockup(
 					property_address,
@@ -203,9 +114,13 @@ const handler = async (
 					block_number.toString(),
 					gasPrice
 				)
+					.on('transactionHash', (hash: string) =>
+						____log('Created the transaction', hash)
+					)
 					.on('confirmation', resolve)
 					.on('error', reject)
 			})
+			____log('Done initilization', property_address, account_address)
 		}
 	)
